RequestHandler - move to Http Action.
